Move health checks and KV ModelView out of core.py and into their own modules
